requirements.txt
fastmcp>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
aiosqlite>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
